+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 #### Improvements
 
 #### Dependency Upgrade
+* Bump Knative Serving to v0.17.2 & Knative Eventing to v0.17.3 
 
 #### New Features
+* Fix #2340: Adding support for Knative Eventing Contrib 
 
 ### 4.11.1 (2020-09-02)